+++ src/gallium/auxiliary/util/u_format_other.c
@@ -26,8 +26,8 @@
  **************************************************************************/
 
 
-#include "u_math.h"
 #include "u_format_other.h"
+#include "util/u_math.h"
 #include "util/format_rgb9e5.h"
 #include "util/format_r11g11b10f.h"
