+++ src/gallium/auxiliary/util/u_format_zs.c
@@ -27,8 +27,8 @@
 
 
 #include "u_debug.h"
-#include "u_math.h"
 #include "u_format_zs.h"
+#include "util/u_math.h"
 
 
 /*
