+++ src/util/u_math.c
